refactor: update path level of base concrete
